added a middleware for the routes
