+++ server/routes/votersRoute.js
@@ -1,23 +1,10 @@
 const express = require("express")
 const router = express.Router()
 const voterModel = require("../models/voterModel")
-const bcrypt = require("bcryptjs")
-const jwt = require("jsonwebtoken")
+const authMiddleware = require("../middlewares/index")
 
 //middleware to getID................................................................
-async function getUser(req, res, next) {
-    let user
-    try {
-        user = await voterModel.findById(req.params.id)
-        if (user == null) {
-            return res.status(404).json({ message: "Cannot find user" })
-        }
-    } catch (error) {
-        return res.status(500).json({ message: error.message })
-    }
-    res.user = user
-    next()
-}
+
 //....................................................................................
 
 
@@ -32,64 +19,13 @@ router.get("/", async (req, res) => {
 })
 
 //get one voter route { by Moderator}
-router.get("/:id", getUser, (req, res) => {
+router.get("/:id", authMiddleware.getUser, (req, res) => {
     res.json(res.user)
 })
 
 //create a voter { SignUp } and { login }
-const registerVoter = (req, res, next) => {
-
-    bcrypt.hash(req.body.password, 5, function (err, hashedPassword) {
-        if (err) {
-            res.json({ error: err })
-        }
-        let voter = new voterModel({
-            name: req.body.name,
-            email: req.body.email,
-            regno: req.body.regno,
-            password: hashedPassword
-        })
-        voter.save()
-            .then(voter => {
-                res.json({ message: "Success! You are added as a voter" })
-            })
-            .catch(error => {
-                res.json({ message: "An err occured!" })
-            })
-    })
-
-
-}
-const loginVoter = (req, res, next) => {
-    let email = req.body.email
-    let password = req.body.password
-
-    voterModel.findOne({ email: email })
-        .then(voter => {
-            if (voter) {
-                bcrypt.compare(password, voter.password, function (err, result) {
-                    if (err) {
-                        res.json({ error: err })
-                    }
-                    if (result) {
-                        let token = jwt.sign({ regno: voter.regno }, "tokenValue", { expiresIn: "1hr" })
-                        res.json({
-                            message: "login successful!",
-                            token: token
-                        })
-                    }
-                    else {
-                        res.json({ message: "Password does not match!" })
-                    }
-                })
-            } else {
-                res.json({ message: "Voter does not exist" })
-            }
-        })
-}
-
-router.post("/reg", registerVoter)
-router.post("/log", loginVoter)
+router.post("/reg", authMiddleware.registerVoter)
+router.post("/log", authMiddleware.loginVoter)
 // router.post("/", registerVoter, async (req, res) => {
 //     try {
 //         const newVoter = await voterModel.create({
@@ -109,7 +45,7 @@ router.post("/log", loginVoter)
 // })
 
 //delete a voter { By Moderator }
-router.delete("/:id", getUser, async (req, res) => {
+router.delete("/:id", authMiddleware.getUser, async (req, res) => {
     try {
         await res.user.remove()
         res.json({ message: "deleted the voter" })
